src/requirements.txt
# Core ML libraries
torch==2.4.0+cu124
torchvision==0.19.0+cu124
# Source: https://download.pytorch.org/whl/cu124
PyYAML==6.0.2